mavlink: move GIMBAL_MANAGER_INFORMATION to separate stream header
